apiBackend/DjangoAPI/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.9
django-cors-headers==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
pip==23.1.2
pymongo==3.12.3
pytz==2023.3
requests==2.31.0
setuptools==67.8.0
simplejson==3.19.1
sqlparse==0.2.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
wheel==0.40.0
